refactor(tests): extract shared SimpleForecaster to openstef_models.testing
- Create packages/openstef-models/src/openstef_models/testing.py with:
  - SimpleForecaster: reusable test stub for the Forecaster interface
  - create_sample_timeseries_dataset(): shared dataset fixture factory
  - create_standard_preprocessing(): shared LagsAdder pipeline factory
- Remove duplicate SimpleForecaster definitions from test_forecasting_model.py
  and test_ensemble_forecasting_model.py; both now import from openstef_models.testing
- Replace duplicate sample_timeseries_dataset fixture bodies with
  one-line calls to create_sample_timeseries_dataset()
- Replace duplicate LagsAdder TransformPipeline blocks in both pickle/
  roundtrip tests with create_standard_preprocessing()
- Fix sonar.cpd.exclusions pattern (examples/tutorials/*.py, not **/py)
  so SonarCloud correctly excludes Jupytext tutorial sources from CPD
