+++ packages/openstef-meta/tests/unit/models/test_ensemble_forecasting_model.py
@@ -4,12 +4,10 @@
 
 import pickle  # noqa: S403 - Controlled test
 from datetime import datetime, timedelta
-from typing import override
+from typing import TYPE_CHECKING, override
 
-import numpy as np
 import pandas as pd
 import pytest
-from pydantic import PrivateAttr
 
 from openstef_core.datasets import ForecastInputDataset
 from openstef_core.datasets.timeseries_dataset import TimeSeriesDataset
@@ -21,42 +19,11 @@ from openstef_core.testing import assert_timeseries_equal, create_synthetic_fore
 from openstef_core.types import LeadTime, Q
 from openstef_meta.models.ensemble_forecasting_model import EnsembleForecastingModel
 from openstef_meta.models.forecast_combiners.forecast_combiner import ForecastCombiner
-from openstef_models.models.forecasting.forecaster import Forecaster
+from openstef_models.testing import SimpleForecaster, create_sample_timeseries_dataset, create_standard_preprocessing
 from openstef_models.transforms.postprocessing.quantile_sorter import QuantileSorter
-from openstef_models.transforms.time_domain.lags_adder import LagsAdder
 
-
-class SimpleForecaster(Forecaster):
-    """Simple test forecaster that returns predictable values for testing."""
-
-    _is_fitted: bool = PrivateAttr(default=False)
-
-    @property
-    @override
-    def hparams(self) -> HyperParams:
-        return HyperParams()
-
-    @property
-    @override
-    def is_fitted(self) -> bool:
-        return self._is_fitted
-
-    @override
-    def fit(self, data: ForecastInputDataset, data_val: ForecastInputDataset | None = None) -> None:
-        self._is_fitted = True
-
-    @override
-    def predict(self, data: ForecastInputDataset) -> ForecastDataset:
-        # Return predictable forecast values
-        forecast_values = {quantile: 100.0 + quantile * 10 for quantile in self.quantiles}
-        return ForecastDataset(
-            pd.DataFrame(
-                {quantile.format(): [forecast_values[quantile]] * len(data.index) for quantile in self.quantiles},
-                index=data.index,
-            ),
-            data.sample_interval,
-            data.forecast_start,
-        )
+if TYPE_CHECKING:
+    from openstef_models.models.forecasting.forecaster import Forecaster
 
 
 class SimpleCombiner(ForecastCombiner):
@@ -116,19 +83,7 @@ class SimpleCombiner(ForecastCombiner):
 @pytest.fixture
 def sample_timeseries_dataset() -> TimeSeriesDataset:
     """Create sample time series data with typical energy forecasting features."""
-    n_samples = 25
-    rng = np.random.default_rng(seed=42)
-
-    data = pd.DataFrame(
-        {
-            "load": 100.0 + rng.normal(10.0, 5.0, n_samples),
-            "temperature": 20.0 + rng.normal(1.0, 0.5, n_samples),
-            "radiation": rng.uniform(0.0, 500.0, n_samples),
-        },
-        index=pd.date_range("2025-01-01 10:00", periods=n_samples, freq="h"),
-    )
-
-    return TimeSeriesDataset(data, timedelta(hours=1))
+    return create_sample_timeseries_dataset()
 
 
 @pytest.fixture
@@ -226,17 +181,7 @@ def test_forecasting_model__pickle_roundtrip():
     original_model = EnsembleForecastingModel(
         forecasters=forecasters,
         combiner=combiner,
-        preprocessing=TransformPipeline(
-            transforms=[
-                LagsAdder(
-                    history_available=timedelta(days=14),
-                    horizons=horizons,
-                    max_day_lags=7,
-                    add_trivial_lags=True,
-                    add_autocorr_lags=False,
-                ),
-            ]
-        ),
+        preprocessing=create_standard_preprocessing(horizons),
         postprocessing=TransformPipeline(transforms=[QuantileSorter()]),
         cutoff_history=timedelta(days=7),
         target_column="load",
