+++ sonar-project.properties
@@ -8,4 +8,4 @@
 # The paired .py jupytext source files share the same boilerplate dataset setup,
 # so they are also excluded for the same reason.
 # Test files are excluded because fixture and helper duplication is expected and intentional.
-sonar.cpd.exclusions=**/*.ipynb,examples/tutorials/**/*.py,**/tests/**
+sonar.cpd.exclusions=**/*.ipynb,examples/tutorials/*.py,**/tests/**/*.py
